Began correction of poor responsive design for new homepage layout and footer; Relocated Radio404.js to Home folder and updated imports, experimenting with track transition sounds
