+++ src/footer/footer.css
@@ -1,3 +1,4 @@
+
 .footer-bg {
     display: flex;
     flex-direction: column;
@@ -8,6 +9,9 @@
     max-width: 100vw;
     background-color: #00BFC3;
     color: white;
+    z-index: 0;
+    margin-top: 100px;
+    padding-bottom: 20px;
 }
 
 .footer-cols {
@@ -92,4 +96,51 @@
 
 .copyright p {
     margin: 0px;
+}
+
+@media (max-width: 790px) {
+    .footer-bg {
+        left: 0px;
+        bottom: 50px;
+        display: flex;
+        flex-direction: row;
+        height: fit-content;
+        justify-content: center;
+    }
+
+    .footer-cols {
+        display: flex;
+        flex-direction: row;
+        flex-wrap: wrap;
+        justify-content: space-evenly;
+        width: fit-content;
+    }
+
+    .footer-links {
+        display: flex;
+        font-size: 15px;
+        justify-content: center;
+    }
+
+    .emailTitle {
+        text-align: center;
+    }
+
+    .emailInput {
+        width: 200px;
+    }
+
+    .mobileHide {
+        display: none;
+    }
+
+    .copyright {
+        position: absolute;
+        text-align: center;
+        max-width: 95%;
+        left: 0px;
+        bottom: 0px;
+        font-size: 15px;
+        margin-left: 10px;
+    }
 }
